Add security to the project and User entity -Add User entity -Add checkCookies method SecurityService class -Add generateCookies method SecurityService class -Add validateUSer method SecurityService class -Rewrite ConnectionFactory implemented DataSource -Add JdbcUserDao class -Add UserDao interface -Add UserRowMapper class -Add UserService class -Add login.html -Add LoginServlet class -Add Flyway configuration to Starter class -Add user_scema.sql -Add V1__Create_product_table.sql for Flyway configuration

+++ src/main/java/web/servlet/DeleteProductServlet.java
@@ -1,7 +1,9 @@
 package web.servlet;
 
 import service.ProductService;
+import service.SecurityService;
 
+import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -11,15 +13,24 @@ import static java.sql.Date.valueOf;
 
 public class DeleteProductServlet extends HttpServlet {
     private final ProductService productService;
+    private final SecurityService securityService;
 
-    public DeleteProductServlet(ProductService productService) {
+    public DeleteProductServlet(ProductService productService, SecurityService securityService) {
         this.productService = productService;
+        this.securityService = securityService;
     }
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        int id = Integer.parseInt(request.getParameter("id"));
-        productService.deleteById(id);
-        response.sendRedirect("/products");
+        Cookie[] cookies = request.getCookies();
+
+        boolean isAuthorized = securityService.checkCookies(cookies);
+        if (isAuthorized) {
+            response.sendRedirect("/login");
+        } else {
+            int id = Integer.parseInt(request.getParameter("id"));
+            productService.deleteById(id);
+            response.sendRedirect("/products");
+        }
     }
 }
